requirements.txt
openai>=1.0.0
moviepy>=1.0.3
python-dotenv>=1.0.0
tenacity>=8.0.0
rich>=13.0.0
pydub>=0.25.1
typing-extensions>=4.0.0